+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `PodCreateOptsBuilder::dns_server`
 * Add `PodCreateOptsBuilder::image_volumes`
 * Add `PodCreateOptsBuilder::mounts`
+* Add `PodCreateOptsBuilder::overlay_volumes`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync